// packages/sdk/src/api/endpoints/entities.ts
export type Media = {
  Account?: {
    avatarUrl: string;
    username: string;
  };
  Game?: any;
  accountId?: string;
  contentType: string;
  cover?: string;
  createdAt?: string;
  description: string;
  gameId?: string | null;
  id?: string;
  name: string;
  updatedAt?: string;
  url: string;
};